+++ R/revdep_over_time.R
@@ -23,19 +23,20 @@ revdep_over_time <- function(pkgs, dates, none = NA_integer_, force = FALSE) {
   cran_revdeps <- import_from("revdepcheck", "cran_revdeps")
   loadCache <- R.cache::loadCache
   saveCache <- R.cache::saveCache
-  
+  dirs <- c(.packageName)
+  dirs_pkg <- c(dirs, "packages")
+
   count_revdeps <- function(pkg) {
-    dirs <- c(.packageName)
     key <- list(
       method = "revdep_over_time",
       pkg = pkg,
       repos = getOption("repos")
     )
-    if (!force && !is.null(n <- loadCache(key, dirs = dirs))) return(n)
+    if (!force && !is.null(n <- loadCache(key, dirs = dirs_pkg))) return(n)
     pkgs <- cran_revdeps(pkg)
     if (length(pkgs) == 0) return(none)
     n <- length(pkgs)
-    saveCache(n, key = key, dirs = dirs)
+    saveCache(n, key = key, dirs = dirs_pkg)
     n
   }
 
@@ -46,15 +47,30 @@ revdep_over_time <- function(pkgs, dates, none = NA_integer_, force = FALSE) {
   p <- progressor(length(dates) * length(pkgs))
   stats <- future_lapply(dates, FUN = function(date) {
     mran_repos <- getSnapshotURL(date, online = FALSE)
-    pmsg <- format(date, format = "%F")
     repos <- c(CRAN = mran_repos)
+
+    key <- list(
+      method = "revdep_over_time",
+      pkgs = pkgs,
+      repos = repos
+    )
+
+    pmsg <- format(date, format = "%F")
+    if (!force && !is.null(stats_pkgs <- loadCache(key, dirs = dirs))) {
+      p(pmsg, amount = length(pkgs))
+      return(stats_pkgs)
+    }
+    
     oopts <- options(repos = repos)
     on.exit(options(oopts), add = TRUE)
-    
-    vapply(pkgs, FUN = function(pkg) {
+    stats_pkgs <- vapply(pkgs, FUN = function(pkg) {
       on.exit(p(pmsg))
       count_revdeps(pkg)
     }, FUN.VALUE = NA_integer_)
+
+    saveCache(stats_pkgs, key = key, dirs = dirs)
+
+    stats_pkgs
   })
 
   stats <- unlist(stats, use.names = FALSE)
@@ -103,12 +119,16 @@ cran_revdep_on_date <- function(package, date, force = FALSE) {
 
 
 getSnapshotURL <- function(date, online = FALSE) {
+  stopifnot(inherits(date, "Date"))
+  
   ## Known time-machine CRAN mirrors
   url_roots <- c(
     MRAN = "https://cran.microsoft.com/snapshot",
     RSPM = "https://packagemanager.rstudio.com/cran"
   )
-  mirror <- getOption("revdepcheck.extras.snapshot.source", "RSPM")
+
+  mirror <- if (date < "2018-01-01") "MRAN" else "RSPM"
+  mirror <- getOption("revdepcheck.extras.snapshot.source", mirror)
   mirror <- match.arg(mirror, choices = names(url_roots))
   url_root <- url_roots[mirror]
   file.path(url_root, date, fsep = "/")