+++ R/run.R
@@ -26,6 +26,8 @@
 #' --preinstall-todo     Pre-install all packages to be checked
 #' --preinstall <pkgs>   Pre-install specified packages
 #' --preinstall-children Install all reverse dependencies
+#' --preinstall-grandchildren
+#'                       Install all second-generation reverse dependencies
 #' --preinstall-update   Install packages that have been updated since last run
 #' --preinstall-error    Install packages that gave an "error" during checks
 #' --preinstall-failure  Install packages that failed to be checked
@@ -226,6 +228,10 @@ run <- function(pkg = ".", ..., warn = 1L, args = base::commandArgs(trailingOnly
     if (identical(pkg, ".")) pkg <- revdep_this_package()
     cran_revdeps <- import_from("revdepcheck", "cran_revdeps")
     pkgs <- cran_revdeps(pkg)
+    revdep_preinstall(pkgs)
+  } else if ("--preinstall-grandchildren" %in% args) {
+    if (identical(pkg, ".")) pkg <- revdep_this_package()
+    pkgs <- revdep_grandchildren(pkg)
     revdep_preinstall(pkgs)
   } else if ("--preinstall-error" %in% args) {
     res <- revdepcheck::revdep_summary(pkg)