+++ NEWS.md
@@ -1,3 +1,8 @@
+# Version (development version)
+
+ * ...
+
+
 # Version 0.5.0 [2023-01-30]
 
 ### New Features