+++ NEWS.md
@@ -1,6 +1,9 @@
 # Version (development version)
 
- * ...
+## New Features
+
+ * Now using [evercran](https://github.com/r-hub/evercran) as the main
+   CRAN snapshot database.
  
 
 # Version 0.6.0 [2023-11-07]