+++ NEWS.md
@@ -2,6 +2,11 @@
 
 ## New Features
 
+ * Added argument `chunk_size` to `revdep_preinstall()`, which
+   controls how many packages are installed at the same time, which in
+   turn affects how frequently **crancache** package library cache is
+   updated.
+
  * Now using [evercran](https://github.com/r-hub/evercran) as the main
    CRAN snapshot database.